netlog visits refac
